chore: ruff configurado e tres desvios da especificacao fechados
Ruff passa limpo em src e tests. As 231 correcoes automaticas foram aplicadas
de uma vez -- uma decisao arriscada -- e verificadas pelo que importa: os 433
testes continuam passando, a linha de comando sobe em 213 ms e o servidor MCP
em 1,0 s sem carregar modulo pesado algum.

As regras desligadas estao justificadas no proprio pyproject: try/except/pass
em caminho de limpeza aqui sempre carrega um comentario dizendo por que a
falha e engolida, e contextlib.suppress esconderia essa explicacao atras de um
nome. O WavFile mantem o arquivo aberto de proposito, por uma reuniao inteira.

Desvios fechados:

- o comando agora se chama `benchmark`, como a especificacao pede, com `bench`
  como apelido para quem digita vinte vezes numa tarde;
- ele aceita um arquivo de configuracoes com rotulo por execucao, entao o
  relatorio nomeia as configuracoes como a pessoa pensa nelas;
- ruff configurado, que faltava na tarefa 1.4.

E um defeito real encontrado ao testar o arquivo de configuracoes: o cache de
modelos cria link simbolico por padrao, e criar um no Windows exige Modo de
Desenvolvedor ou elevacao. Baixar um modelo falhava com "o cliente nao tem o
privilegio necessario" numa maquina onde tudo o mais funciona. Copiar custa
disco e sempre da certo, que e a troca certa para um cache.

+++ voxvault-core/src/voxvault/engine/media.py
@@ -83,7 +83,7 @@ def require_ffmpeg() -> str:
 def already_normalized(path: Path) -> bool:
     """True when the file is already 16 kHz mono 16-bit PCM."""
     try:
-        import soundfile as sf  # noqa: PLC0415  (lazy: keeps CLI startup cheap)
+        import soundfile as sf
 
         info = sf.info(str(path))
     except Exception:
@@ -98,11 +98,11 @@ def already_normalized(path: Path) -> bool:
 def probe_duration_ms(path: Path) -> int | None:
     """Duration in milliseconds, or None when it cannot be determined."""
     with contextlib.suppress(Exception):
-        import soundfile as sf  # noqa: PLC0415
+        import soundfile as sf
 
         info = sf.info(str(path))
         if info.samplerate:
-            return int(round(info.frames / info.samplerate * 1000))
+            return round(info.frames / info.samplerate * 1000)
 
     exe = find_ffprobe()
     if not exe:
@@ -119,7 +119,7 @@ def probe_duration_ms(path: Path) -> int | None:
     if out.returncode != 0 or not text:
         return None
     try:
-        return int(round(float(text) * 1000))
+        return round(float(text) * 1000)
     except ValueError:
         return None
 
@@ -207,8 +207,8 @@ def verify_lossless(original: Path, compressed: Path) -> bool:
     unverified conversion is not a recoverable mistake.
     """
     try:
-        import numpy as np  # noqa: PLC0415
-        import soundfile as sf  # noqa: PLC0415
+        import numpy as np
+        import soundfile as sf
 
         a, rate_a = sf.read(str(original), dtype="int16", always_2d=True)
         b, rate_b = sf.read(str(compressed), dtype="int16", always_2d=True)
